backend/requirements.txt
elasticsearch==8.15.1
fastapi==0.115.0
pillow==10.4.0 
python-dotenv==1.0.1
sentence-transformers==3.1.1
wptools==0.4.17
uvicorn==0.31.0